+++ example/lib/models/magazine.freezed.dart
@@ -20,9 +20,7 @@ mixin _$Magazine {
   String get author => throw _privateConstructorUsedError;
   int get year => throw _privateConstructorUsedError;
 
-  /// Create a copy of Magazine
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $MagazineCopyWith<Magazine> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -45,8 +43,6 @@ class _$MagazineCopyWithImpl<$Res, $Val extends Magazine>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of Magazine
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -90,8 +86,6 @@ class __$$MagazineImplCopyWithImpl<$Res>
       _$MagazineImpl _value, $Res Function(_$MagazineImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of Magazine
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -147,9 +141,7 @@ class _$MagazineImpl implements _Magazine {
   @override
   int get hashCode => Object.hash(runtimeType, title, author, year);
 
-  /// Create a copy of Magazine
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$MagazineImplCopyWith<_$MagazineImpl> get copyWith =>
@@ -168,11 +160,8 @@ abstract class _Magazine implements Magazine {
   String get author;
   @override
   int get year;
-
-  /// Create a copy of Magazine
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$MagazineImplCopyWith<_$MagazineImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }