+++ publish.ps1
@@ -2,7 +2,7 @@
 
 # 1. Generate version
 Write-Host "==> Generating version from pubspec.yaml..."
-dart run tool/generate_version.dart
+dart run bin/generate_version.dart
 if ($LASTEXITCODE -ne 0) {
     Write-Error "Error while generating version."
     exit 1
@@ -10,7 +10,7 @@ if ($LASTEXITCODE -ne 0) {
 
 # 2. Format the generated file
 Write-Host "==> Formatting version.g.dart..."
-dart format lib/src/version.g.dart
+dart format lib/src/generated/version.g.dart
 
 # 3. Publish the package
 Write-Host "==> Publishing package to pub.dev..."